feat(hex): real corpus_fs CorpusReader/Writer adapter + wire consumers (soc-vzyhn #corpus-fs-adapter)
Extract the FS-backed BC1 Corpus adapters out of package main into a real
cli/internal/adapters/corpus_fs/ package, mirroring the merged storage_fs /
workspace_git / tracker_bd siblings. The hex audit flagged the corpus FS
(.agents/ knowledge store) as called directly from cli/cmd/ao without living
behind a proper adapter package.

- corpus_fs.Reader implements ports.CorpusReaderPort (root-scoped *.md walk
  via os.OpenRoot, symlink-escape rejection, title/body substring ranker).
- corpus_fs.Writer implements ports.CorpusWriterPort (scoped, idempotent
  write with sorted YAML-frontmatter rendering; sentinel errors
  ErrRootRequired / ErrPathRequired / ErrPathEscapesRoot).
- Compile-time `var _ ports.CorpusReaderPort/CorpusWriterPort` checks.
- Wired both consumers: `ao corpus inject` -> corpus_fs.NewReader,
  `ao corpus capture` -> corpus_fs.NewWriter. Removed the in-package
  productionCorpusReader/Writer types + tests they replaced.
- Real temp-dir FS tests (no live ao/bd dependency); errors.Is on sentinels.

The in-memory doubles (ports.InMemoryCorpus{Reader,Writer}) already exist and
stay green. Follow-up: no other direct .agents/ corpus callsites in cli/cmd/ao
remained for these two ports; broader corpus-FS callsite migration (mine/forge
decay-ranker paths) is out of scope for this slice.
